feat: ScheduleRepository 추가
야호

+++ lipit_app/app/src/main/java/com/ssafy/lipit_app/data/model/response_dto/schedule/ScheduleResponse.kt
@@ -1,7 +1,7 @@
 package com.ssafy.lipit_app.data.model.response_dto.schedule
 
 data class ScheduleResponse(
-    val callScheduleId: Int,
+    val callScheduleId: Long,
     val scheduledDay: String,
     val scheduledTime: String,
     val topicCategory: String
